Fix: compilation error in idt

+++ src/idt.c
@@ -1,5 +1,7 @@
 #include "idt.h"
 
+// https://wiki.osdev.org/Interrupt_Descriptor_Table
+
 IDT32 idt[IDT_ENTRY];
 
 static void create_table(u32 offset, u16 selector, u8 type, u32 entry) {
@@ -15,6 +17,6 @@ void init_idt() {
 	idt_ptr->limit = (sizeof(IDT32) * IDT_ENTRY) - 1;
 	idt_ptr->base = (u32)&idt;
 
-	load_idt(idt_ptr);
-    return ;
+	load_idt((u32)idt_ptr);
+	return ;
 }
